acteur repository debut method

+++ src/main/java/projet/ais/repository/ContinentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Continent;
 
-public interface ContinentRepository  extends JpaRepository<Continent , Integer>{
+public interface ContinentRepository  extends JpaRepository<Continent , Long>{
     
 }
