+++ src/main/java/projet/ais/repository/MagasinRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Magasin;
 
-public interface MagasinRepository  extends JpaRepository<Magasin, Integer>{
+public interface MagasinRepository  extends JpaRepository<Magasin, Long>{
     
 }
