+++ src/main/java/projet/ais/repository/Niveau1PaysRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Niveau1Pays;
 
-public interface Niveau1PaysRepository extends JpaRepository<Niveau1Pays, Integer>{
+public interface Niveau1PaysRepository extends JpaRepository<Niveau1Pays, Long>{
     
 }
