+++ src/main/java/projet/ais/repository/Niveau2PaysRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Niveau2Pays;
 
-public interface Niveau2PaysRepository  extends JpaRepository<Niveau2Pays, Integer>{
+public interface Niveau2PaysRepository  extends JpaRepository<Niveau2Pays, Long>{
     
 }
