+++ src/main/java/projet/ais/repository/PaysRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Pays;
 
-public interface PaysRepository  extends JpaRepository<Pays,Integer>{
+public interface PaysRepository  extends JpaRepository<Pays,Long>{
     
 }
