+++ src/main/java/projet/ais/repository/RegroupementParametreRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.RegroupementParametre;
 
-public interface RegroupementParametreRepository  extends JpaRepository<RegroupementParametre, Integer>{
+public interface RegroupementParametreRepository  extends JpaRepository<RegroupementParametre, Long>{
     
 }
