+++ src/main/java/projet/ais/repository/Sortie_StockRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.Sortie_Stock;
 
-public interface Sortie_StockRepository extends JpaRepository<Sortie_Stock, Integer>{
+public interface Sortie_StockRepository extends JpaRepository<Sortie_Stock, Long>{
     
 }
