+++ src/main/java/projet/ais/repository/TypeActeurRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.TypeActeur;
 
-public interface TypeActeurRepository extends JpaRepository<TypeActeur, Integer>{
+public interface TypeActeurRepository extends JpaRepository<TypeActeur, Long>{
     
 }
