+++ src/main/java/projet/ais/repository/ZoneProductionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import projet.ais.models.ZoneProduction;
 
-public interface ZoneProductionRepository extends JpaRepository<ZoneProduction, Integer>{
+public interface ZoneProductionRepository extends JpaRepository<ZoneProduction, Long>{
     
 }
